OboeTester: factor out analyzer
So we can use it in CTS or other projects.
