+++ apps/OboeTester/app/src/main/cpp/FullDuplexGlitches.h
@@ -22,6 +22,7 @@
 
 #include "oboe/Oboe.h"
 #include "FullDuplexAnalyzer.h"
+#include "analyzer/GlitchAnalyzer.h"
 
 class FullDuplexGlitches : public FullDuplexAnalyzer {
 public:
